added state in address table an views

+++ controller/dbUpdater.js
@@ -6,14 +6,31 @@ const path = require('path');
 let dbStorageDir = __dirname  + path.sep + ['..', 'db_storage', ''].join(path.sep);
 
 function createBackup() {
-    createStorageDir();
-    let currentDate = new Date();
-    let dateVersion = currentDate.toISOString().replace(/-/g, '').replace(/:/g, '').replace('.','');
-    fs.copyFileSync(__dirname + path.sep + '../seelsorge.db', dbStorageDir + 'seelsorgeBackup_'+dateVersion+'.db');
+    let existsStorageDir = createStorageDir();
+    if (existsStorageDir) {
+        let currentDate = new Date();
+        let dateVersion = currentDate.toISOString().replace(/-/g, '').replace(/:/g, '').replace('.','');
+        fs.copyFileSync(__dirname + path.sep + '../seelsorge.db', dbStorageDir + 'seelsorgeBackup_'+dateVersion+'.db');
+    }
 }
 
 function checkIsNewVersionOnline() {
-    console.log("download new file");
+    const Client = require('ssh2-sftp-client');
+    let sftp = new Client();
+    const sftpConfig = require('../config/sftp.json');
+
+    sftp.connect({
+        host: sftpConfig.host,
+        port: sftpConfig.port,
+        username: sftpConfig.user,
+        password: sftpConfig.password
+      }).then(() => {
+        return sftp.list('/pathname');
+      }).then(data => {
+        console.log(data, 'the data info');
+      }).catch(err => {
+        console.log(err, 'catch error');
+      });
 }
 
 function createStorageDir() {
@@ -21,15 +38,17 @@ function createStorageDir() {
     if (!fs.existsSync(dirPath)) {
         fs.mkdir(dirPath, (err) => {
             if (err) {
-                return console.error(err);
+                return false;
             }
             fs.chmod(dirPath, 0o777, (err) => {
-                if (err) throw err;
-                console.log('The permissions set for dir: db_storage');
+                if (err) 
+                    return false;
               });
-            console.log(dirPath + ' Directory created successfully!');
+            return true;
         });
     }
+
+    return true;
 }
 
 module.exports =  {
